src/utils/ocr/extract/pokemon.test.ts
import {describe, expect, it} from '@jest/globals';

import {ocrExtractPokemonInfo} from '@/utils/ocr/extract/pokemon';

/* eslint-disable max-len */
const ocrResult = `
-~ €re 505 S A ) - Lv.5 Eevee — ®Lv.30 @ Ly.60 - , :‘ [,-'”‘}} "r'q; B S o G @ T o Every 1 hr 1 min 10 secs e 12 | Main skill & Sub skills Ingredient Magnet S Lv.1 Gets you 6 ingredients chosen at random. [ O Lv.25 Berry Finding S Helping Bonus (o .. co SRR o ;S Helping Speed M Inventory Up M Ingredient Finder M I Additional Stats o Main skill chance A A Sassy EXP gains VV - (=] ‘ i G LAMERS
7
`;
/* eslint-enable max-len */


describe('OCR / Extract Pokemon Info', () => {
  it('extracts correctly', () => {
    const extracted = ocrExtractPokemonInfo({
      text: ocrResult,
      translations: {
        name: {
          'Eevee': 5,
        },
        subSkill: {
          'Berry Finding S': 7,
          'Helping Bonus': 8,
          'Helping Speed M': 9,
          'Inventory Up M': 10,
          'Ingredient Finder M': 11,
        },
        nature: {
          'Sassy': 3,
        },
      },
    });

    expect(extracted.pokemonId).toBe(5);
    expect(extracted.subSkills[0].id).toBe(7);
    expect(extracted.subSkills[0].level).toBe(10);
    expect(extracted.subSkills[1].id).toBe(8);
    expect(extracted.subSkills[1].level).toBe(25);
    expect(extracted.subSkills[2].id).toBe(9);
    expect(extracted.subSkills[2].level).toBe(50);
    expect(extracted.subSkills[3].id).toBe(10);
    expect(extracted.subSkills[3].level).toBe(75);
    expect(extracted.subSkills[4].id).toBe(11);
    expect(extracted.subSkills[4].level).toBe(100);
    expect(extracted.nature).toBe(3);
  });
});
